move to pages folder
