+++ src/pages/notFoundPage/NotFoundPage.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { withRouter } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
-import Button from '../../../components/input/Button';
-import { LANDING_PAGE_ROUTE } from '../../../Routes';
+import Button from '../../components/input/Button';
+import { LANDING_PAGE_ROUTE } from '../../Routes';
 
 import { NotFoundPageWrapper, NotFoundImage, PageHeader, HeaderText } from './styles/NotFoundPage';
